Recreated the inheriteance tree for the Flags to allow for better polymorphism. Types now follow the inheritance tree that their names would suggest. Also altered the directory names
