+++ include/Flag.h
@@ -2,6 +2,7 @@
 #define CMD_LINE_PARSER_FLAGS_H
 
 #include <string>
+#include <string_view>
 #include <vector>
 #include <stdexcept>
 #include <iostream>
@@ -20,9 +21,6 @@ namespace CmdLineParser
 
 	template<typename T>
 	concept StringType = std::same_as<T, std::string>;
-
-	template<typename T>
-	concept FlagType = std::is_base_of<Flag, T>::value;
 
 	class Flag
 	{
@@ -114,116 +112,6 @@ namespace CmdLineParser
 		Flag(const Flag&) = delete;
 
 		Flag& operator=(const Flag&) = delete;
-	};
-
-	class BranchTrigger : public Flag
-	{
-	protected:
-		const flag_event* _triggeredFunc;
-		std::vector<Flag> _nestedFlags;
-
-	public:
-		BranchTrigger(std::string&& flagToken, std::string&& flagDesc,
-			bool flagRequired = false)
-#ifndef _DEBUG
-			noexcept
-#endif // !_DEBUG
-			;
-
-		template<FlagType... Flags>
-		BranchTrigger(std::string&& flagToken, std::string&& flagDesc, const flag_event& triggeredFunc, Flags&&... subFlags,
-			bool flagRequired = false) 
-#ifndef _DEBUG
-			noexcept
-#endif // !_DEBUG
-			: Flag(std::move(flagToken), std::move(flagDesc), flagRequired), _triggeredFunc(&triggeredFunc), _nestedFlags{std::forward<Flags>(subFlags)...}
-		{
-#ifdef _DEBUG
-			if (flagToken.length() == 1)
-				throw std::invalid_argument("Error: A Branch Flag should have one recognizable token that is a long token, not a single character short token");
-#endif // _DEBUG
-		}
-
-		template<FlagType... Flags>
-		BranchTrigger(std::string&& flagToken, std::string&& flagDesc, const flag_event& triggeredFunc, const flag_argument& flagArg, Flags&&... subFlags,
-			bool argRequired = false, bool flagRequired = false) 
-#ifndef _DEBUG
-			noexcept
-#endif // !_DEBUG
-			: Flag(std::move(flagToken), std::move(flagDesc), flagArg, argRequired, flagRequired), _triggeredFunc(&triggeredFunc), _nestedFlags{ std::forward<Flags>(subFlags)... }
-		{
-#ifdef _DEBUG
-			if (flagToken.length() == 1)
-				throw std::invalid_argument("Error: A Branch Flag should have one recognizable token that is a long token, not a single character short token");
-#endif // _DEBUG
-		}
-
-		template<FlagType... Flags>
-		BranchTrigger& SetSubFlags(Flags&&... subFlags) noexcept
-		{
-			_nestedFlags.clear();
-			(_nestedFlags.emplace_back(std::forward<Flags>(subFlags)), ...);
-
-			return *this;
-		}
-
-		BranchTrigger& SetFlagEvent(const flag_event& triggeredFunc) noexcept;
-
-		void Raise(std::vector<std::string_view>::const_iterator& itr, const std::vector<std::string_view>::const_iterator end) override;
-
-		bool TryRaise(std::vector<std::string_view>::const_iterator& itr, const std::vector<std::string_view>::const_iterator end, std::string* errorMsg = nullptr) noexcept override;
-
-		BranchTrigger(const BranchTrigger&) = delete;
-		BranchTrigger(BranchTrigger&&) = delete;
-
-		BranchTrigger& operator=(const BranchTrigger&) = delete;
-		BranchTrigger& operator=(BranchTrigger&&) = delete;
-	};
-
-	class BranchTriggerSwitch : public SwitchFlag
-	{
-	protected:
-		const flag_event* _triggeredFunc;
-		std::vector<Flag> _nestedFlags;
-
-	public:
-		BranchTriggerSwitch(std::string&& flagToken, std::string&& flagDesc,
-			bool defaultSwitchState = false, bool flagRequired = false)
-#ifndef _DEBUG
-			noexcept
-#endif // !_DEBUG
-			;
-
-		template<FlagType... Flags>
-		BranchTriggerSwitch(std::string&& flagToken, std::string&& flagDesc, const flag_event& triggeredFunc, Flags&&... subFlags,
-			bool defaultSwitchState = false, bool flagRequired = false) 
-#ifndef _DEBUG
-			noexcept
-#endif // !_DEBUG
-			: SwitchFlag(std::move(flagToken), std::move(flagDesc), defaultSwitchState, flagRequired), _triggeredFunc(&triggeredFunc), _nestedFlags{ std::forward<Flags>(subFlags)... }
-		{
-		}
-
-		template<FlagType... Flags>
-		BranchTriggerSwitch& SetSubFlags(Flags&&... subFlags) noexcept
-		{
-			_nestedFlags.clear();
-			(_nestedFlags.emplace_back(std::forward<Flags>(subFlags)), ...);
-
-			return *this;
-		}
-
-		BranchTriggerSwitch& SetFlagEvent(const flag_event& triggeredFunc) noexcept;
-
-		void Raise(std::vector<std::string_view>::const_iterator& itr, const std::vector<std::string_view>::const_iterator end) override;
-
-		bool TryRaise(std::vector<std::string_view>::const_iterator& itr, const std::vector<std::string_view>::const_iterator end, std::string* errorMsg = nullptr) noexcept override;
-
-		BranchTriggerSwitch(const BranchTriggerSwitch&) = delete;
-		BranchTriggerSwitch(BranchTriggerSwitch&&) = delete;
-
-		BranchTriggerSwitch& operator=(const BranchTriggerSwitch&) = delete;
-		BranchTriggerSwitch& operator=(BranchTriggerSwitch&&) = delete;
 	};
 	//************************************************************************************************
