+++ include/FlagDerivedTypes/TriggerFlag.h
@@ -5,12 +5,19 @@
 
 namespace CmdLineParser
 {
-	class TriggerFlag : public Flag
+	class TriggerFlag : virtual public Flag
 	{
 	protected:
 		const flag_event* _triggeredFunc;
 
 	public:
+		TriggerFlag(std::string&& flagToken, std::string&& flagDesc,
+			bool flagRequired = false)
+#ifndef _DEBUG
+			noexcept
+#endif // !_DEBUG
+			;
+
 		template<StringType... Tokens>
 		TriggerFlag(Tokens&&... flagTokens, std::string&& flagDesc,
 			bool flagRequired = false)
@@ -20,6 +27,13 @@ namespace CmdLineParser
 			: Flag(std::forward<Tokens>(flagTokens)..., std::move(flagDesc), flagRequired)
 		{}
 
+		TriggerFlag(std::string&& flagToken, std::string&& flagDesc, const flag_event& triggeredFunc,
+			bool flagRequired = false)
+#ifndef _DEBUG
+			noexcept
+#endif // !_DEBUG
+			;
+
 		template<StringType... Tokens>
 		TriggerFlag(Tokens&&... flagTokens, std::string&& flagDesc, const flag_event& triggeredFunc,
 			bool flagRequired = false)
@@ -28,6 +42,13 @@ namespace CmdLineParser
 #endif // !_DEBUG
 			: Flag(std::forward<Tokens>(flagTokens)..., std::move(flagDesc), flagRequired), _triggeredFunc(&triggeredFunc)
 		{}
+		
+		TriggerFlag(std::string&& flagToken, std::string&& flagDesc, const flag_event& triggeredFunc, const flag_argument& flagArg,
+			bool argRequired = false, bool flagRequired = false)
+#ifndef _DEBUG
+			noexcept
+#endif // !_DEBUG
+			;
 
 		template<StringType... Tokens>
 		TriggerFlag(Tokens&&... flagTokens, std::string&& flagDesc, const flag_event& triggeredFunc, const flag_argument& flagArg,
